requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
flask_bcrypt
email_validator
python-decouple
gunicorn
